let child tool implemenet the craft method
