+++ src/items/tools/knife.js
@@ -1,4 +1,7 @@
 import Tool from "./tool.js";
+import { getItemClass } from "../../utilities.js";
+import { characterData } from "../../cookie-io.js";
+import { CONSTANTS } from "../../constants/constants.js";
 
 export default class Knife extends Tool {
     // Item data
@@ -36,5 +39,46 @@ export default class Knife extends Tool {
         }
 
         return output;
+    }
+
+    async craft(item) {
+        let outputString = "";
+        const recipe = this.getRecipe(item.name);
+        console.log("Combining", this.name, item.name);
+
+        // Fletch item if possible
+        if (item.numItems >= recipe.numRequiredItems) {
+            const dashboard = characterData.getScene(CONSTANTS.SCENES.DASHBOARD);
+
+            if (recipe.className) {
+                let newItem = await getItemClass(recipe.className, dashboard);
+
+                // Get name before adding it to inventory because
+                // if it's a duplicate it will be destroyed
+                const newItemName = newItem.name;
+
+                // Item was added
+                if (dashboard.inventory.addToInventory(newItem)) {
+                    item.setNumItems(item.numItems - recipe.numRequiredItems);
+
+                    outputString = "Fletched " + newItemName + ".";
+                    characterData.addSkillXp({ fletching: recipe.xpGiven });
+                }
+            }
+            // Insufficient materials
+            else if (item.numItems < recipe.numRequiredItems) {
+                outputString =
+                    recipe.numRequiredItems +
+                    " " +
+                    item.name +
+                    " are needed to fletch that";
+            }
+        } else {
+            outputString = "Not a valid fletching combination.";
+        }
+
+        // Write to chat window
+        const chatScene = characterData.getScene(CONSTANTS.SCENES.CHAT);
+        chatScene.writeText(outputString);
     }
 }
