+++ src/items/tools/tinderbox.js
@@ -1,4 +1,6 @@
 import Tool from "./tool.js";
+import { characterData } from "../../cookie-io.js";
+import { CONSTANTS } from "../../constants/constants.js";
 
 export default class Tinderbox extends Tool {
     // Item data
@@ -17,7 +19,7 @@ export default class Tinderbox extends Tool {
 
     getRecipe(itemName) {
         let output = {
-            className: "",
+            className: null,
             numRequiredItems: 0,
             xpGiven: 0,
         };
@@ -34,5 +36,32 @@ export default class Tinderbox extends Tool {
         }
 
         return output;
+    }
+
+    async craft(item) {
+        let outputString = "";
+        const recipe = this.getRecipe(item.name);
+        console.log("Combining", this.name, item.name);
+
+        // Burn log if possible
+        if (item.numItems >= recipe.numRequiredItems) {
+            item.setNumItems(item.numItems - recipe.numRequiredItems);
+            outputString = "Firemaked " + item.name + ".";
+            characterData.addSkillXp({ firemaking: recipe.xpGiven });
+        }
+        // Insufficient materials
+        else if (item.numItems < recipe.numRequiredItems) {
+            outputString =
+                recipe.numRequiredItems +
+                " " +
+                item.name +
+                " are needed to firemake that that";
+        } else {
+            outputString = "Not a valid firemaking combination.";
+        }
+
+        // Write to chat window
+        const chatScene = characterData.getScene(CONSTANTS.SCENES.CHAT);
+        chatScene.writeText(outputString);
     }
 }
