+++ src/targets/furnace.js
@@ -45,10 +45,12 @@ export class Furnace extends ClickableObject {
         const selectedIndex = inv.curSelectedItemIndex;
         const selectedItem = inv.inventory[selectedIndex];
 
+        console.log(selectedItem);
+
         if (
             selectedIndex < 0 ||
             !selectedItem ||
-            !this.validMaterials.has(selectedItem)
+            !this.validMaterials.has(selectedItem.name)
         ) {
             chat.writeText("Select an ore in your inventory first.");
             return;
@@ -87,7 +89,7 @@ export class Furnace extends ClickableObject {
 
                 // Log click for stats
                 this.scene.stats.updateClickedTargetStat();
-                characterData.addSkillXp({smithing: bar.xp});
+                characterData.addSkillXp({ smithing: bar.xp });
                 this.scene.enemyKilled("bar");
 
                 // Show smelt animation
